# CY4/mongodb/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Biblioteca API",
        "description": "Essa é a documentação para a API de um banco de dados de uma biblioteca."
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
        "/books":{
            "post": {
                "summary": "Cria um novo livro",
                "description": "Adiciona um novo livro na biblioteca. O novo livro deve conter todos os parâmetros obrigatórios",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Dados do livro que será adicionado",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "title": {
                                    "type": "string",
                                    "example": "A Divina Comédia"
                                },
                                "author": {
                                    "type": "string",
                                    "example": "Dante"
                                },
                                "year": {
                                    "type": "integer",
                                    "example": 1321
                                },
                                "genre": {
                                    "type": "string",
                                    "example": "poesia"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Livro adicionado com sucesso!"
                    },
                    "500": {
                        "description": "Erro ao adicionar livro"
                    }
                }
            },
            "get": {
                "summary": "Mostra as informações do livro em questão",
                "description": "Mostra todos os dados de um livro presente na biblioteca através de seu ID.",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Filtrar por título do livro",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "author",
                        "in": "query",
                        "description": "Filtrar por autor do livro",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Livro não encontrado"
                    },
                    "200": {
                        "description": "Lista de livros",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "id": { "type": "string", "example": "68b628788cf2ccd91071c38c" },
                                    "title": { "type": "string", "example":"A Divina Comédia" },
                                    "author": { "type":"string", "example": "Dante" },
                                    "year": { "type":"integer", "example":1321 },
                                    "genre": { "type":"string", "example":"poesia" }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao buscar livros"
                    }
                }
            }
        },
        "/books/{id}":{
            "patch": {
                "summary": "Atualiza um livro da biblioteca",
                "description": "As informações preenchidas devem conter todos os parâmetros obrigatórios de um livro",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do livro que será atualizado",
                        "required": true,
                        "type": "string",
                        "example": "68b628788cf2ccd91071c38c"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Dados do livro que será atualizado",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "title": {
                                    "type": "string",
                                    "example": "A Divina Comédia"
                                },
                                "author": {
                                    "type": "string",
                                    "example": "Dante"
                                },
                                "year": {
                                    "type": "integer",
                                    "example": 1321
                                },
                                "genre": {
                                    "type": "string",
                                    "example": "poesia"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Livro atualizado com sucesso"
                    },
                    "404": {
                        "description": "Livro não encontrado"
                    },
                    "500": {
                        "description": "Erro ao atualizar o livro"
                    }
                }
            },
            "delete": {
                "summary": "Deleta um livro da biblioteca",
                "description": "Apaga um livro escolhido da biblioteca através de uma sequência de letras e números que serão chamados de seu ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do livro que será deletado",
                        "required": true,
                        "type": "string",
                        "example": "68b628788cf2ccd91071c38c"
                    }
                ],
                "responses": {
                    "404": {
                        "description":"Livro não encontrado" 
                    },
                    "201": {
                        "description": "Livro excluído com sucesso"
                    },
                    "500":{
                        "description": "Erro ao tentar excluir o livro"

                    }
                }
            }
        }
    }
}